Make Reserve component DRY
